(feature) Worker to log requests
This is a very simple worker that asynchronously logs requests into the
database, so they can be floded up into the format `Transition` expectes
later.

NB: We **could** have done this by writing Redis keys, but based on the
volume or traffic we anticiapte for this, this should be sufficient for
now.