(feature) Log all requests asynchronously
There are some issues to resolve here, we use a catch-all to prevent the
logger erroring, if you remove this, a lot of tests fail because it
depends on too many things in the stack due to some demeter violations.
This needs resolving somehow.

Otherwise this works very well and can handle quite a bit of load.